+++ src/test/java/page_object/HomePageScooter.java
@@ -1,4 +1,4 @@
-package PageObject;
+package page_object;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
