+++ src/test/java/page_object/constants/Stands.java
@@ -1,0 +1,5 @@
+package page_object.constants;
+
+public class Stands {
+    public static final String SCOOTER_SERVICES = "https://qa-scooter.praktikum-services.ru/";
+}
